+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - Modules [8FACE, 8FACEx2](./docs/EightFace.md)
     - Load preset in Arm-mode even when the same slot was selected before (#212)
+- Module [MIDI-CAT](./docs/MidiCat.md)
+    - Added support for MIDI 14-bit CC
 
 ## 1.8.0